+++ src/lib/i18n/locales/en-GB/translation.json
@@ -505,6 +505,7 @@
 	"Import Prompts": "",
 	"Import Tools": "",
 	"Import V1 Prompts": "",
+	"Incident Form": "",
 	"Include": "",
 	"Include `--api-auth` flag when running stable-diffusion-webui": "",
 	"Include `--api` flag when running stable-diffusion-webui": "",
@@ -517,9 +518,6 @@
 	"Invalid file format.": "",
 	"Invalid Tag": "",
 	"is typing...": "",
-	"Issue": "",
-	"Issue and Suggestion Form": "",
-	"Issue Form": "",
 	"January": "",
 	"Jina API Key": "",
 	"join our Discord for help.": "",
@@ -709,14 +707,12 @@
 	"Playground": "",
 	"Please carefully review the following warnings:": "",
 	"Please describe what happened": "",
-	"Please describe your suggestion in detail": "",
 	"Please enter a prompt": "",
 	"Please fill in all fields.": "",
 	"Please fill out all required fields": "",
 	"Please list the steps to reproduce this issue": "",
 	"Please select a model first.": "",
 	"Please select a reason": "",
-	"Please select an item in the list": "",
 	"Port": "",
 	"Positive attitude": "",
 	"Prefix ID": "",
@@ -760,7 +756,7 @@
 	"Reorder Models": "",
 	"Repeat Last N": "",
 	"Reply in Thread": "",
-	"Report an Issue": "",
+	"Report an Incident": "",
 	"Report any problems or bugs you encounter to help us improve the application.": "",
 	"Request Mode": "",
 	"Reranking Model": "",
@@ -821,7 +817,6 @@
 	"Select a pipeline": "",
 	"Select a pipeline url": "",
 	"Select a tool": "",
-	"Select a type...": "",
 	"Select an engine": "",
 	"Select an Ollama instance": "",
 	"Select Engine": "",
@@ -864,8 +859,6 @@
 	"Share": "",
 	"Share Chat": "",
 	"Share to OpenWebUI Community": "",
-	"Share your feedback to help us improve - report issues or suggest new features.": "",
-	"Share your ideas for new features or improvements to make the application better.": "",
 	"Show": "",
 	"Show \"What's New\" modal on login": "",
 	"Show Admin Details in Account Pending Overlay": "",
@@ -890,14 +883,11 @@
 	"Stream Chat Response": "",
 	"STT Model": "",
 	"STT Settings": "",
-	"Submit Report": "",
-	"Submitting...": "",
+	"Submit Incident": "",
 	"Subtitle (e.g. about the Roman Empire)": "",
 	"Success": "",
 	"Successfully updated.": "",
 	"Suggested": "",
-	"Suggestion": "",
-	"Suggestion Form": "",
 	"Support": "",
 	"Support this plugin:": "",
 	"Survey": "",
