+++ src/lib/i18n/locales/fr-CA/translation.json
@@ -505,6 +505,7 @@
 	"Import Prompts": "Importer des invites",
 	"Import Tools": "Outils d'importation",
 	"Import V1 Prompts": "Importer les invites V1",
+	"Incident Form": "",
 	"Include": "Inclure",
 	"Include `--api-auth` flag when running stable-diffusion-webui": "Inclure le drapeau `--api-auth` lors de l'exécution de stable-diffusion-webui",
 	"Include `--api` flag when running stable-diffusion-webui": "Inclure le drapeau `--api` lorsque vous exécutez stable-diffusion-webui",
@@ -517,9 +518,6 @@
 	"Invalid file format.": "Format de fichier invalide.",
 	"Invalid Tag": "Étiquette non valide",
 	"is typing...": "est en train d'écrire...",
-	"Issue": "Problème",
-	"Issue and Suggestion Form": "Formulaire de problème et de suggestion",
-	"Issue Form": "Formulaire de problème",
 	"January": "Janvier",
 	"Jina API Key": "Clé API Jina",
 	"join our Discord for help.": "Rejoignez notre Discord pour obtenir de l'aide.",
@@ -709,14 +707,12 @@
 	"Playground": "Aire de jeux",
 	"Please carefully review the following warnings:": "Veuillez examiner attentivement les avertissements suivants :",
 	"Please describe what happened": "Veuillez décrire ce qui s'est passé",
-	"Please describe your suggestion in detail": "Veuillez décrire votre suggestion en détail",
 	"Please enter a prompt": "Veuillez entrer une invite",
 	"Please fill in all fields.": "Veuillez remplir tous les champs.",
 	"Please fill out all required fields": "Veuillez remplir tous les champs obligatoires",
 	"Please list the steps to reproduce this issue": "Veuillez énumérer les étapes pour reproduire ce problème",
 	"Please select a model first.": "Veuillez d'abord sélectionner un modèle.",
 	"Please select a reason": "Veuillez sélectionner une raison",
-	"Please select an item in the list": "Veuillez sélectionner un élément dans la liste",
 	"Port": "Port",
 	"Positive attitude": "Attitude positive",
 	"Prefix ID": "ID préfixe",
@@ -760,7 +756,7 @@
 	"Reorder Models": "Réorganiser les modèles",
 	"Repeat Last N": "Répéter les N derniers",
 	"Reply in Thread": "Répondre dans le fil",
-	"Report an Issue": "Signaler un problème",
+	"Report an Incident": "",
 	"Report any problems or bugs you encounter to help us improve the application.": "Signalez tout problème ou bogue que vous rencontrez pour nous aider à améliorer l'application.",
 	"Request Mode": "Mode de Requête",
 	"Reranking Model": "Modèle de reclassement",
@@ -821,7 +817,6 @@
 	"Select a pipeline": "Sélectionnez un pipeline",
 	"Select a pipeline url": "Sélectionnez l'URL du pipeline",
 	"Select a tool": "Sélectionnez un outil",
-	"Select a type...": "Sélectionner un genre...",
 	"Select an engine": "Sélectionnez un moteur",
 	"Select an Ollama instance": "Sélectionnez une instance Ollama",
 	"Select Engine": "Sélectionnez Moteur",
@@ -864,8 +859,6 @@
 	"Share": "Partager",
 	"Share Chat": "Partage de chat",
 	"Share to OpenWebUI Community": "Partager avec la communauté OpenWebUI",
-	"Share your feedback to help us improve - report issues or suggest new features.": "Partagez vos commentaires pour nous aider à nous améliorer - signalez des problèmes ou suggérez de nouvelles fonctionnalités.",
-	"Share your ideas for new features or improvements to make the application better.": "Partagez vos idées de nouvelles fonctionnalités ou d'améliorations pour rendre l'application meilleure.",
 	"Show": "Montrer",
 	"Show \"What's New\" modal on login": "Afficher la fenêtre modale « Quoi de neuf » lors de la connexion",
 	"Show Admin Details in Account Pending Overlay": "Afficher les détails de l'administrateur dans la superposition en attente du compte",
@@ -890,14 +883,11 @@
 	"Stream Chat Response": "Diffuser la réponse du chat",
 	"STT Model": "Modèle de reconnaissance vocale",
 	"STT Settings": "Paramètres de reconnaissance vocale",
-	"Submit Report": "Soumettre le rapport",
-	"Submitting...": "Soumission en cours...",
+	"Submit Incident": "Soumettre l'incident",
 	"Subtitle (e.g. about the Roman Empire)": "Sous-titres (par ex. sur l'Empire romain)",
 	"Success": "Réussite",
 	"Successfully updated.": "Mise à jour réussie.",
 	"Suggested": "Suggéré",
-	"Suggestion": "Suggestion",
-	"Suggestion Form": "Formulaire de suggestion",
 	"Support": "Soutien",
 	"Support this plugin:": "Soutenez ce plugin :",
 	"Survey": "Sondage",
